Give show_doc its own command class
